requirements.txt
Flask
torch
torchvision
matplotlib
Pillow